+++ .gitignore
@@ -1,1 +1,1 @@
-virtualenv
+*.pyc
