+++ producer/requirements.txt
@@ -1,0 +1,6 @@
+gevent==1.3.7
+greenlet==0.4.15
+six==1.11.0
+websocket-client==0.54.0
+eventlet==0.24.1
+flake8==3.7.7
