+++ producer/sallibs.config
@@ -1,0 +1,13 @@
+SALPY_ATArchiver
+SALPY_ATCamera
+SALPY_ATHeaderService
+SALPY_ATMonochromator
+SALPY_ATPtg
+SALPY_Electrometer
+SALPY_Environment
+SALPY_FiberSpectrograph
+SALPY_LinearStage
+SALPY_Script
+SALPY_ScriptQueue
+SALPY_Test
+SALPY_TunableLaser
