+++ producer/start-daemon.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+source .setup.sh
+
+python -u /usr/src/love/main.py
