--- telemetry-producer/producer/start-daemon.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-source .setup.sh
-
-python -u /usr/src/love/main.py
