--- telemetry-producer/producer/utils.py
@@ -1,10 +1,0 @@
-import numpy as np
-import json
-
-class NumpyEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, np.ndarray):
-            return obj.tolist()
-        if isinstance(obj, np.int32):
-            return int(obj)
-        return json.JSONEncoder.default(self, obj)
